Add files via upload
In my project which is job portal candidates can see the Latest jobs as well as search for the job by designation as well as location. To apply for the particular job candidates first need to log in. An employer can post the job after subscribing to the package. In this project, I have developed one module in such a way that employees can send messages to candidates after looking at their resumes. An employee can able to download the resume of the candidates. Candidates receive messages from employees on the notification icon. Both Candidates and Employees were able to edit their Information. On the website, I am using Forget password concept where clicking on the forget password link URL sends to their Gmail account, and the candidate is able to reset their password. (My job portal Website is hosted on PythonAnywhere you can check it out.)
Link: https://newJobportal.pythonanywhere.com/